+++ tapo/examples/tapo_p300.rs
@@ -2,7 +2,7 @@
 use std::{env, thread, time::Duration};
 
 use log::{info, LevelFilter};
-use tapo::{ApiClient, PlugIdentifier};
+use tapo::{ApiClient, Plug};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -35,9 +35,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             child.nickname, child.device_id, child.device_on,
         );
 
-        let plug = power_strip
-            .plug(PlugIdentifier::ByDeviceId(&child.device_id))
-            .await?;
+        let plug = power_strip.plug(Plug::ByDeviceId(&child.device_id)).await?;
 
         info!("Turning device on...");
         plug.on().await?;